+++ erpnext/accounts/doctype/sales_invoice/sales_invoice.py
@@ -330,7 +330,6 @@ class SalesInvoice(SellingController):
 				against_acc.append(d.income_account)
 		self.against_income_account = ','.join(against_acc)
 
-
 	def add_remarks(self):
 		if not self.remarks: self.remarks = 'No Remarks'
