+++ erpnext/demo/data/item_schools.json
@@ -1,0 +1,94 @@
+[
+ {
+  "default_supplier": "Asiatic Solutions",
+  "default_warehouse": "Stores",
+  "item_code": "Books",
+  "item_group": "Raw Material",
+  "item_name": "Books"
+ },
+ {
+  "default_supplier": "HomeBase",
+  "default_warehouse": "Stores",
+  "item_code": "Pencil",
+  "item_group": "Raw Material",
+  "item_name": "Pencil"
+ },
+ {
+  "default_supplier": "New World Realty",
+  "default_warehouse": "Stores",
+  "item_code": "Tables",
+  "item_group": "Raw Material",
+  "item_name": "Tables"
+ },
+ {
+  "default_supplier": "Eagle Hardware",
+  "default_warehouse": "Stores",
+  "item_code": "Chair",
+  "item_group": "Raw Material",
+  "item_name": "Chair"
+ },
+ {
+  "default_supplier": "Asiatic Solutions",
+  "default_warehouse": "Stores",
+  "item_code": "Black Board",
+  "item_group": "Sub Assemblies",
+  "item_name": "Black Board"
+ },
+ {
+  "default_supplier": "HomeBase",
+  "default_warehouse": "Stores",
+  "item_code": "Chalk",
+  "item_group": "Raw Material",
+  "item_name": "Chalk"
+ },
+ {
+  "default_supplier": "HomeBase",
+  "default_warehouse": "Stores",
+  "item_code": "Notepad",
+  "item_group": "Raw Material",
+  "item_name": "Notepad"
+ },
+ {
+  "default_supplier": "Ks Merchandise",
+  "default_warehouse": "Stores",
+  "item_code": "Uniform",
+  "item_group": "Raw Material",
+  "item_name": "Uniform"
+ },
+ {
+  "is_stock_item": 0,
+  "description": "Computer",
+  "item_code": "Computer",
+  "item_name": "Computer",
+  "item_group": "Products"
+ },
+ {
+  "is_stock_item": 0,
+  "description": "Mobile",
+  "item_code": "Mobile",
+  "item_name": "Mobile",
+  "item_group": "Products"
+ },
+ {
+  "is_stock_item": 0,
+  "description": "ERP",
+  "item_code": "ERP",
+  "item_name": "ERP",
+  "item_group": "All Item Groups"
+ },
+ {
+  "is_stock_item": 0,
+  "description": "Autocad",
+  "item_code": "Autocad",
+  "item_name": "Autocad",
+  "item_group": "All Item Groups"
+ },
+ {
+  "default_warehouse": "Stores",
+  "item_code": "Service",
+  "item_group": "Services",
+  "item_name": "Service",
+  "has_variants": 0,
+  "is_stock_item": 0
+ }
+]
