+++ erpnext/demo/data/program.json
@@ -1,0 +1,102 @@
+[
+	{
+		"doctype": "Program",
+		"program_name": "Masters of Computer Applications",
+		"program_code": "MCA",
+		"department": "Information Technology",
+		"courses": [
+			{
+				"course": "Microprocessor",
+				"academic_term": "Semester 1"
+			},
+			{
+				"course": "Probability and Statistics",
+				"academic_term": "Semester 1"
+			},
+			{
+				"course": "Programing in Java",
+				"academic_term": "Semester 2"
+			}
+		]
+	},
+	{
+		"doctype": "Program",
+		"program_name": "Bachelor of Computer Applications",
+		"program_code": "BCA",
+		"department": "Information Technology",
+		"courses": [
+			{
+				"course": "Communication Skiils",
+				"academic_term": "Semester 3"
+			},
+			{
+				"course": "Object Oriented Programing - C++",
+				"academic_term": "Semester 3"
+			},
+			{
+				"course": "Basic Mathematics",
+				"academic_term": "Semester 2"
+			},
+			{
+				"course": "Data Structures and Algorithm",
+				"academic_term": "Semester 2"
+			},
+			{
+				"course": "Digital Logic",
+				"academic_term": "Semester 2"
+			},
+			{
+				"course": "Fundamentals of IT & Programing",
+				"academic_term": "Semester 1"
+			},
+			{
+				"course": "Operating System",
+				"academic_term": "Semester 1"
+			},
+			{
+				"course": "Programing in C",
+				"academic_term": "Semester 1"
+			}
+		]
+	},
+	{
+		"doctype": "Program",
+		"program_name": "Bachelor of Business Administration",
+		"program_code": "BBA",
+		"department": "Management Studies",
+		"courses": [
+			{
+				"course": "Organizational Behavior",
+				"academic_term": "Semester 1"
+			},
+			{
+				"course": "Management Development and Skills",
+				"academic_term": "Semester 1"
+			},
+			{
+				"course": "Legal and Regulatory Framework",
+				"academic_term": "Semester 1"
+			},
+			{
+				"course": "Human Resource Management",
+				"academic_term": "Semester 2"
+			},
+			{
+				"course": "Entrepreneurship Management",
+				"academic_term": "Semester 2"
+			},
+			{
+				"course": "Communication Skills",
+				"academic_term": "Semester 2"
+			},
+			{
+				"course": "Business Environment",
+				"academic_term": "Semester 3"
+			},
+			{
+				"course": "Advertising and Sales",
+				"academic_term": "Semester 3"
+			}
+		]
+	}
+]
