+++ erpnext/demo/data/user.json
@@ -26,8 +26,8 @@
  },
  {
   "email": "aromn@example.com",
-  "first_name": "\u0414\u043c\u0438\u0442\u0440\u0438\u0439",
-  "last_name": "\u041f\u0438\u0440\u043e\u0433\u043e\u0432"
+  "first_name": "Arom",
+  "last_name": "Nolan"
  },
  {
   "email": "TildeLindqvist@example.com",
