+++ erpnext/demo/demo.py
@@ -3,9 +3,8 @@ from __future__ import unicode_literals
 import frappe, sys
 import erpnext
 import frappe.utils
-from erpnext.demo.setup_data import setup_data
-from erpnext.demo.user import hr, sales, purchase, manufacturing, stock, accounts, projects, fixed_asset
-
+from erpnext.demo.user import hr, sales, purchase, manufacturing, stock, accounts, projects, fixed_asset, schools
+from erpnext.demo.setup import education, manufacture, setup_data
 """
 Make a demo
 
@@ -26,14 +25,19 @@ bench --site demo.erpnext.dev execute erpnext.demo.demo.simulate
 def make(domain='Manufacturing'):
 	frappe.flags.domain = domain
 	frappe.flags.mute_emails = True
-	setup_data()
+	setup_data.setup(domain)
+	if domain== 'Manufacturing':
+		manufacture.setup_data()
+	elif domain== 'Education':
+		education.setup_data()
+	
 	site = frappe.local.site
 	frappe.destroy()
 	frappe.init(site)
 	frappe.connect()
-	simulate()
+	simulate(domain)
 
-def simulate():
+def simulate(domain='Manufacturing'):
 	runs_for = frappe.flags.runs_for or 150
 	frappe.flags.company = erpnext.get_default_company()
 	frappe.flags.mute_emails = True
@@ -47,33 +51,38 @@ def simulate():
 	# continue?
 	demo_last_date = frappe.db.get_global('demo_last_date')
 	if demo_last_date:
-		current_date = frappe.utils.add_days(demo_last_date, 1)
-
+		current_date = frappe.utils.add_days(frappe.utils.getdate(demo_last_date), 1)
+		
 	# run till today
 	if not runs_for:
 		runs_for = frappe.utils.date_diff(frappe.utils.nowdate(), current_date)
 		# runs_for = 100
 
 	fixed_asset.work()
-
 	for i in xrange(runs_for):
 		sys.stdout.write("\rSimulating {0}".format(current_date.strftime("%Y-%m-%d")))
 		sys.stdout.flush()
 		frappe.flags.current_date = current_date
-
 		if current_date.weekday() in (5, 6):
 			current_date = frappe.utils.add_days(current_date, 1)
 			continue
-
-		hr.work()
-		sales.work()
-		purchase.work()
-		manufacturing.work()
-		stock.work()
-		accounts.work()
-		projects.run_projects(current_date)
-		# run_messages()
-
-		current_date = frappe.utils.add_days(current_date, 1)
-
-		frappe.db.commit()
+		try:
+			hr.work()
+			purchase.work()
+			stock.work()
+			accounts.work()
+			projects.run_projects(current_date)
+			#run_messages()
+			
+			if domain=='Manufacturing':
+				sales.work()
+				manufacturing.work()
+			elif domain=='Education':
+				schools.work()
+		
+		except:
+			frappe.db.set_global('demo_last_date', current_date)
+			raise	
+		finally:
+			current_date = frappe.utils.add_days(current_date, 1)
+			frappe.db.commit()
