+++ erpnext/demo/user/projects.py
@@ -5,8 +5,7 @@ from __future__ import unicode_literals
 
 import frappe
 from frappe.utils import flt
-from frappe.utils.make_random import can_make
-from frappe.utils.make_random import how_many, get_random
+from frappe.utils.make_random import get_random
 from erpnext.projects.doctype.timesheet.test_timesheet import make_timesheet
 from erpnext.demo.user.hr import make_sales_invoice_for_timesheet
 
@@ -20,13 +19,12 @@ def run_projects(current_date):
 def make_timesheet_for_projects(current_date	):
 	for data in frappe.get_all("Task", ["name", "project"], {"status": "Open", "exp_end_date": ("<", current_date)}):
 		employee = get_random("Employee")
-		if frappe.db.get_value('Salary Structure', {'employee': employee}, 'salary_slip_based_on_timesheet'):
-			ts = make_timesheet(employee, simulate = True, billable = 1,
-				activity_type=get_random("Activity Type"), project=data.project, task =data.name)
+		ts = make_timesheet(employee, simulate = True, billable = 1,
+			activity_type=get_random("Activity Type"), project=data.project, task =data.name)
 
-			if flt(ts.total_billing_amount) > 0.0:
-				make_sales_invoice_for_timesheet(ts.name)
-				frappe.db.commit()
+		if flt(ts.total_billing_amount) > 0.0:
+			make_sales_invoice_for_timesheet(ts.name)
+			frappe.db.commit()
 
 def close_tasks(current_date):
 	for task in frappe.get_all("Task", ["name"], {"status": "Open", "exp_end_date": ("<", current_date)}):
