+++ erpnext/demo/user/stock.py
@@ -102,23 +102,21 @@ def submit_draft_stock_entries():
 def make_sales_return_records():
 	for data in frappe.get_all('Delivery Note', fields=["name"], filters={"docstatus": 1}):
 		if random.random() < 0.2:
-			print "dn"
 			try:
 				dn = make_sales_return(data.name)
 				dn.insert()
 				dn.submit()
 				frappe.db.commit()
-			except Exception, e:
+			except Exception:
 				frappe.db.rollback()
 
 def make_purchase_return_records():
 	for data in frappe.get_all('Purchase Receipt', fields=["name"], filters={"docstatus": 1}):
 		if random.random() < 0.2:
-			print "purchase"
 			try:
 				pr = make_purchase_return(data.name)
 				pr.insert()
 				pr.submit()
 				frappe.db.commit()
-			except Exception, e:
+			except Exception:
 				frappe.db.rollback()
